% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{barplot_bivar}
\alias{barplot_bivar}
\title{Easy Barplot}
\usage{
barplot_bivar(tab, x, y, graphPercent = NULL, showGraphNA = NULL)
}
\arguments{
\item{tab}{a data frame}

\item{x}{the independant variable}

\item{y}{the dependant variable}
}
\value{
a ggplot barplot
}
\description{
Easy Barplot
}
