% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{barplot_desc}
\alias{barplot_desc}
\title{Easy barplot}
\usage{
barplot_desc(
  tab,
  x,
  ylab = gettext("proportion"),
  showGraphNA = NULL,
  palette = "hue",
  graphPercent = NULL
)
}
\arguments{
\item{tab}{a data frame}

\item{x}{the variable to display}
}
\value{
a ggplot barplot
}
\description{
Easy barplot
}
