% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Calculate AUC}
\usage{
calculate_auc(pred_obs)
}
\arguments{
\item{mod}{The prediction model}

\item{orig_table}{The original tab}

\item{vardep}{The dependant varible}
}
\value{

}
\description{
Calculate AUC
}
