% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{compute_cv_perf}
\alias{compute_cv_perf}
\title{Get performance measures of each subsample among a n-fold cross validation}
\usage{
compute_cv_perf(tab, vardep, varindep, type, n = 10)
}
\arguments{
\item{tab}{The data.frame}

\item{vardep}{The dependant variable}

\item{varindep}{The independant variable, forced into the model}

\item{type}{A character vector, one of "linear", "logistic" or "survival}

\item{n}{n-fold cross validation}
}
\value{
The mean AUC
}
\description{
Get performance measures of each subsample among a n-fold cross validation
}
