% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-desc.R
\name{create_ligne_desc}
\alias{create_ligne_desc}
\alias{create_ligne_desc.numeric}
\alias{create_ligne_desc.factor}
\alias{create_ligne_desc_export}
\alias{create_ligne_desc_export.factor}
\alias{create_ligne_desc_export.default}
\title{Descriptive analysis}
\usage{
create_ligne_desc(x, ...)

\method{create_ligne_desc}{numeric}(x, noms, ...)

\method{create_ligne_desc}{factor}(x, noms, ...)

create_ligne_desc_export(x, ...)

\method{create_ligne_desc_export}{factor}(x, noms, ...)

create_ligne_desc_export.default(x, ...)
}
\arguments{
\item{x}{A vector to describe}

\item{noms}{variable names}
}
\value{
If x is a numeric vector: mean(sd), median(Q25-75), min, max, n.
    If x is a factor: n (%)
}
\description{
Descriptive analysis
}
