% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-bivar.R
\name{create_ligne_surv_bivar}
\alias{create_ligne_surv_bivar}
\title{Univariate analysis of survival}
\usage{
create_ligne_surv_bivar(x, time, noms, censure)
}
\arguments{
\item{x}{The independant variable}

\item{time}{The vector of follow-up time}

\item{censure}{The status indicator}
}
\value{
median(IC95), max follow-up, n, n events, p.
}
\description{
Univariate analysis of survival
}
