% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-bivar.R
\name{create_line_bivar}
\alias{create_line_bivar}
\alias{create_line_bivar.factor}
\alias{create_line_bivar.numeric}
\title{Univariate analysis}
\usage{
create_line_bivar(x, ...)

\method{create_line_bivar}{factor}(x, y, noms, margin = 2, .drop = TRUE, compute_p = TRUE, ...)

\method{create_line_bivar}{numeric}(x, y, noms, .drop = TRUE, compute_p = TRUE, summary = NULL)
}
\arguments{
\item{x}{The independant variable}

\item{y}{A vector to describe}

\item{noms}{The variable names}

\item{margin}{Index, or vector of indices to generate margin for.
1 indicates rows, 2 indicates columns}

\item{compute_p}{logical: should the function compute the pvalue}
}
\value{
If x is a numeric vector: mean(sd), median(Q25-75), min, max, n, p.
    If x is a factor: n (%)
}
\description{
Univariate analysis
}
