% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-desc.R
\name{create_line_desc}
\alias{create_line_desc}
\alias{create_line_desc.numeric}
\alias{create_line_desc.factor}
\alias{create_line_desc_export}
\alias{create_line_desc_export.factor}
\alias{create_line_desc_export.numeric}
\title{Descriptive analysis}
\usage{
create_line_desc(x, ...)

\method{create_line_desc}{numeric}(x, noms, ...)

\method{create_line_desc}{factor}(x, noms, show_prop = TRUE)

create_line_desc_export(x, ...)

\method{create_line_desc_export}{factor}(x, noms, show_prop = TRUE)

\method{create_line_desc_export}{numeric}(x, noms, ...)
}
\arguments{
\item{x}{A vector to describe}

\item{noms}{variable names}
}
\value{
If x is a numeric vector: mean(sd), median(Q25-75), min, max, n.
    If x is a factor: n (%)
}
\description{
Descriptive analysis
}
