% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-desc.R
\name{create_line_surv_desc}
\alias{create_line_surv_desc}
\title{Descriptive analysis of survival}
\usage{
create_line_surv_desc(time, censure)
}
\arguments{
\item{time}{The vector of follow-up time}

\item{censure}{The status indicator}
}
\value{
median(IC95), max follow-up, n, n events.
}
\description{
Descriptive analysis of survival
}
