% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{create_pred_obs}
\alias{create_pred_obs}
\title{Creates a contingency table with predicted probabilities and actual classification}
\usage{
create_pred_obs(
  mod,
  tab = NULL,
  vardep = NULL,
  as_prob = TRUE,
  shrunk = numeric(0)
)
}
\arguments{
\item{mod}{a glm or mira object}
}
\value{
data.frame
}
\description{
Creates a contingency table with predicted probabilities and actual classification
}
