% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tab.R
\name{create_tabi}
\alias{create_tabi}
\title{Create intermediate table}
\usage{
create_tabi(tab, type, keep = NULL)
}
\arguments{
\item{tab}{a data frame}

\item{type}{a character vector. Possible values : "desc", "expl" and "pred"}
}
\value{
a curated data frame
}
\description{
Create intermediate table
}
