% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustments.R
\name{define_varAjust}
\alias{define_varAjust}
\title{Detect adjustment variables}
\usage{
define_varAjust(tab, vardep, varindep, type, test = FALSE)
}
\arguments{
\item{tab}{A data frame}

\item{vardep}{A character string of the dependant variable}

\item{varindep}{A character vector of the independant variables}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}
}
\value{

}
\description{
Detect adjustment variables
}
