% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{display_missing}
\alias{display_missing}
\title{This code is a modification of the ggmice package - https://github.com/amices/ggmice)}
\usage{
display_missing(data, square = TRUE, rotate = TRUE)
}
\description{
This code is a modification of the ggmice package - https://github.com/amices/ggmice)
}
