% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_from_model}
\alias{extract_from_model}
\title{Extract objects from a model}
\usage{
extract_from_model(model, vector)
}
\arguments{
\item{model}{an object of type "lm" or "coxph"}

\item{vector}{A character vector of object to extract}
}
\value{
A vector
}
\description{
Extract objects from a model
}
