% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pval.R
\name{extract_pval}
\alias{extract_pval}
\title{Performs univariate tests and extract p-value from objects}
\usage{
extract_pval(x, y, survival = FALSE, censure = NULL)
}
\arguments{
\item{x}{The dependant variable}

\item{y}{Either the independant variable if survival = FALSE or follow-up time}

\item{survival}{logical, is it a survival analysis}

\item{censure}{The censor variable}
}
\description{
Performs univariate tests and extract p-value from objects
}
