% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-multiv.R
\name{find_best_precision}
\alias{find_best_precision}
\title{Get the best precision for all indepedant variables}
\usage{
find_best_precision(tab, varindep)
}
\arguments{
\item{tab}{The data.frame}

\item{varindep}{The independant variables}
}
\value{
a character vector
}
\description{
Get the best precision for all indepedant variables
}
