% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{find_varaux}
\alias{find_varaux}
\title{Find auxillary variables, ie predictors of data missingness}
\usage{
find_varaux(
  tab,
  vardep,
  varindep = character(0),
  varajust = character(0),
  type
)
}
\arguments{
\item{tab}{A data.frame}

\item{vardep}{The dependant variable}

\item{varindep}{A character vector of independant variables}

\item{varajust}{A character vector of covariates}

\item{type}{Can be one of "linear", "logistic", "survival"}
}
\description{
Find auxillary variables, ie predictors of data missingness
}
