% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_number}
\alias{format_number}
\title{Format numbers}
\usage{
format_number(numbers, digits = 3)
}
\arguments{
\item{numbers}{a numeric vector}

\item{digits}{number of digits}
}
\value{
a character vector
}
\description{
Format numbers
}
