% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_confint_p_boot}
\alias{get_confint_p_boot}
\title{Extract CI and p-value from bootstraps}
\usage{
get_confint_p_boot(resBoot, resBoot_p)
}
\arguments{
\item{resBoot}{an object of type boot returning coef}

\item{resBoot_p}{an object of type boot returning the test statistic}
}
\value{
A 2-items list:
  type_ci: indicating the type of Confidence Interval computed
  tabRet: the data frame
}
\description{
Extract CI and p-value from bootstraps
}
