% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-multiv.R
\name{get_fun}
\alias{get_fun}
\title{Get adequate function and arguments for model}
\usage{
get_fun(type)
}
\arguments{
\item{type}{a character vector being one of "linear", "logistic" or "survival"}
}
\value{
a list
}
\description{
Get adequate function and arguments for model
}
