% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{get_glance}
\alias{get_glance}
\title{Get additional info on a test performed}
\usage{
get_glance(x, y, survival = FALSE, censure = NULL)
}
\arguments{
\item{x}{The dependant variable}

\item{y}{Either the independant variable if survival = FALSE or follow-up time}

\item{survival}{logical, is it a survival analysis}

\item{censure}{The censor variable}
}
\description{
Get additional info on a test performed
}
