% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustments.R
\name{get_lasso_variables}
\alias{get_lasso_variables}
\title{Variable selection by LASSO procedure}
\usage{
get_lasso_variables(
  tab,
  vardep,
  varindep = character(0),
  type = "logistic",
  sparse = TRUE
)

get_lasso_variables(
  tab,
  vardep,
  varindep = character(0),
  type = "logistic",
  sparse = TRUE
)
}
\arguments{
\item{tab}{A data.frame}

\item{vardep}{The dependant variable}

\item{varindep}{The independant variables to force into the model}

\item{type}{Type of analysis: linear, logistic or survival}

\item{sparse}{logical: should the variable selection be parsimonious}
}
\value{
A character vector of the predictor variables

A character vector of the predictor variables
}
\description{
Variable selection by LASSO procedure

Variable selection by LASSO procedure
}
