% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_pred_perf}
\alias{get_pred_perf}
\title{Get performance measures of a predictive model}
\usage{
get_pred_perf(
  tab,
  vardep,
  varindep = NULL,
  type = "logistic",
  type_validation = "cv",
  R = 100,
  nCPU = 1L,
  mod = NULL,
  updateProgress = function(detail) detail
)
}
\arguments{
\item{tab}{The data.frame}

\item{vardep}{The dependant variable}

\item{varindep}{The independant variable, forced into the model}

\item{type}{A character vector, one of "linear", "logistic" or "survival}

\item{type_validation}{character "cv" or "boot": either performs bootstrap, or cross validation}

\item{R}{The number of bootstrap replicates}

\item{nCPU}{integer: number of processes to be used in parallel operation}
}
\value{
a list containing The mean AUC if cross validation; if bootstrap: the confidence interval and excess optimism
}
\description{
Get performance measures of a predictive model
}
