% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_resBoot}
\alias{get_resBoot}
\alias{get_resBoot_p}
\alias{get_resBoot_anova}
\title{Perform bootstraps}
\usage{
get_resBoot(
  tab,
  R,
  model_base,
  nCPU,
  updateProgress = function(detail) detail,
  var_ajust = NULL
)

get_resBoot_p(
  tab,
  R,
  model_base,
  nCPU,
  updateProgress = function(detail) detail,
  var_ajust = NULL
)

get_resBoot_anova(
  tab,
  R,
  model_base,
  nCPU,
  updateProgress = function(detail) detail,
  var_ajust = NULL
)
}
\arguments{
\item{tab}{A tab}

\item{R}{Number of replications of the bootstrap}

\item{model_base}{The base model}

\item{nCPU}{The number of CPU used in parallel processing}

\item{updateProgress}{Function to show when used in Shiny}

\item{var_ajust}{The adjustment variables of the model}
}
\value{
An object of class boot
}
\description{
Perform bootstraps
}
\section{Functions}{
\itemize{
\item \code{get_resBoot}: Performs bootstrap to further compute the confidence interval

\item \code{get_resBoot_p}: Permutation test for p-values

\item \code{get_resBoot_anova}: Permutation test for anova p-values
}}

\examples{


}
