% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{histogram}
\alias{histogram}
\title{Easy histogram}
\usage{
histogram(tab, x, barlength, ylab)
}
\arguments{
\item{tab}{a data frame}

\item{x}{the variable to display}
}
\value{
a ggplot histogram
}
\description{
Easy histogram
}
