% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputer}
\alias{imputer}
\title{Imputation strategy}
\usage{
imputer(tab, type, vardep, n_imputation = 1)
}
\arguments{
\item{tab}{A data frame}

\item{type}{Can be one of "linear", "logistic", "survival"}

\item{vardep}{The dependant variable}

\item{n_imputation}{Number of imputations}
}
\value{
Either a data frame or a S3 object of class mids
}
\description{
Imputation strategy
}
