% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_error}
\alias{is_error}
\alias{is_warning}
\title{Helper function to get error or warning class}
\usage{
is_error(x)

is_warning(x)
}
\arguments{
\item{x}{a variable}
}
\description{
Helper function to get error or warning class
}
