% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{is_homoscedatic}
\alias{is_homoscedatic}
\alias{is_homoscedatic.lm}
\alias{is_homoscedatic.default}
\title{Can it be considered homoscedatic?}
\usage{
is_homoscedatic(x, ...)

\method{is_homoscedatic}{lm}(x, ...)

\method{is_homoscedatic}{default}(x, y, ...)
}
\arguments{
\item{x}{Object to verify}

\item{...}{other arguments passed to the generic function}

\item{y}{The grouping variable}

\item{model}{The result of the lm function}
}
\description{
Can it be considered homoscedatic?
}
