% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tab.R
\name{make_tab_survival}
\alias{make_tab_survival}
\title{Add column .time}
\usage{
make_tab_survival(
  tab,
  vardep,
  passage = 1,
  typeCensure = 2,
  dateInclusion = NULL,
  dateSortie = NULL,
  var_time = NULL,
  limit = NULL
)
}
\arguments{
\item{tab}{The data frame}

\item{vardep}{The dependent variable}

\item{passage}{If 1, The first level of the event variable is the absence of event; if 2, the second level is the absence of event}

\item{typeCensure}{If 1: Follow-up time computed from 2 dates: inclusion and censor; \
If 2: the follow-up time is already inside a column}

\item{dateInclusion}{Date vector of length equals to the number of rows of tab: inclusion date.}

\item{dateSortie}{Date vector of length equals to the number of rows of tab: censor or event.}

\item{var_time}{Name of the column indicating the follow-up time follow-up}
}
\value{
A table with .time column
}
\description{
Adds the column .time, which is the follow-up time
}
\details{
This function requires either var_time or dateSortie and dateInclusion
}
