% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest_up_thousand}
\alias{nearest_up_thousand}
\title{Get the nearest (ceiling) thousand}
\usage{
nearest_up_thousand(x)
}
\arguments{
\item{x}{a length one numeric vector}
}
\value{
the nearest thousand
}
\description{
Get the nearest (ceiling) thousand
}
