% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical-checks.R
\name{plot_all_splines}
\alias{plot_all_splines}
\title{Plots splines}
\usage{
plot_all_splines(tab, vardep, varindep, varajust, type)
}
\arguments{
\item{tab}{The data frame}

\item{vardep}{A character string of dependent variable}

\item{varindep}{A character vector of independant variables}

\item{varajust}{A character vector of adjustment variables}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}

\item{pred}{A logical indicating whether the analysis is predictive}
}
\value{
All plots
}
\description{
Plots splines
}
