% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical-checks.R
\name{plot_all_zph}
\alias{plot_all_zph}
\title{Title}
\usage{
plot_all_zph(tab, vardep, varindep, var_ajust)
}
\arguments{
\item{tab}{The data frame}

\item{vardep}{A character string of dependent variable}

\item{varindep}{A character vector of independant variables}

\item{var_ajust}{A character vector of adjustment variables}
}
\value{
All plots
}
\description{
Title
}
