% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_reglin}
\alias{plot_reglin}
\title{Plots the relationship between 2 quantitative variables}
\usage{
plot_reglin(tab, x, y, method = "lm")
}
\arguments{
\item{tab}{a data frame}

\item{x}{the independant variable}

\item{y}{the dependant variable}
}
\value{
a ggplot plot
}
\description{
Plots the relationship between 2 quantitative variables
}
