% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print_all_boots}
\alias{print_all_boots}
\title{Print all bootstraps}
\usage{
print_all_boots(model_base, var_ajust = NULL, nCPU = 2, R = NULL)
}
\arguments{
\item{model_base}{The base model}

\item{var_ajust}{The adjustment variables of the model}

\item{nCPU}{The number of CPU used in parallel processing}

\item{R}{Number of replications of the bootstraps}
}
\value{
A data frame with coefficients, p-values and global p-values (computed by ANOVA)
}
\description{
Computes all necessary bootstraps
}
