% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-bivar.R
\name{print_line_bivar}
\alias{print_line_bivar}
\title{Displays the univariate analysis in markdown}
\usage{
print_line_bivar(x, y, varname = label(x), censure = NULL, type = "linear")
}
\arguments{
\item{x}{The independant variable}

\item{y}{The variable to describe or the follow-up time for survival analysis}

\item{censure}{The status indicator}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}
}
\value{
A tibble displayed in markdown
}
\description{
Displays the univariate analysis in markdown
}
