% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-desc.R
\name{print_line_desc}
\alias{print_line_desc}
\title{Displays the descriptive analysis in markdown}
\usage{
print_line_desc(x, varname = NA, type = "linear", time = NULL)
}
\arguments{
\item{x}{The variable to describe}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}

\item{time}{The vector of follow-up time}
}
\value{
A tibble displayed in markdown
}
\description{
Displays the descriptive analysis in markdown
}
