% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{print_plot_bivar}
\alias{print_plot_bivar}
\title{Plot univariate graphs}
\usage{
print_plot_bivar(tab, vardep, varindep, type = "linear")
}
\arguments{
\item{tab}{The data frame}

\item{vardep}{The dependent variable}

\item{varindep}{The independant variable}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}
}
\value{
a ggplot2 graph
}
\description{
Plot univariate graphs
}
