% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{print_plot_desc}
\alias{print_plot_desc}
\title{Plot descriptive graphs}
\usage{
print_plot_desc(tab, vardep, var_n, type = "linear")
}
\arguments{
\item{tab}{The data frame}

\item{vardep}{The dependent variable}

\item{var_n}{The independant variable}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}
}
\value{
a ggplot2 graph
}
\description{
Plot descriptive graphs
}
