% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_tab_export}
\alias{read_tab_export}
\title{Read file}
\usage{
read_tab_export(file, sep = ",", dec = ".")
}
\arguments{
\item{file}{The file to read}

\item{sep}{The field separator character. Values on each line of the file are separated by this character. If sep = "" (the default for read.table) the separator is ‘white space’, that is one or more spaces, tabs, newlines or carriage returns.}

\item{dec}{The character used in the file for decimal points.}
}
\value{
A data.frame
}
\description{
This is a wrapper of read_tab_imp
}
