% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_tab_import}
\alias{read_tab_import}
\title{Read file}
\usage{
read_tab_import(file, sep = "\\t", dec = ".")
}
\arguments{
\item{file}{The file to read}

\item{sep}{The field separator character. Values on each line of the file are separated by this character.}

\item{dec}{The character used in the file for decimal points.}
}
\value{
A data.frame
}
\description{
import txt, csv and xls files
}
