% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustments.R
\name{recherche_multicol}
\alias{recherche_multicol}
\title{Search for multicolinearity}
\usage{
recherche_multicol(tab, vardep, varindep, var_ajust, type)
}
\arguments{
\item{tab}{A data frame}

\item{vardep}{A character string of the dependant variable}

\item{varindep}{A character vector of the independant variables}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}

\item{varAjust}{A character vector of the adjustment variables}
}
\value{
A character vector of all terms with a VIF > 5
}
\description{
This function removes all terms of the model with a VIF > 5
}
