% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_table_markdown}
\alias{show_table_markdown}
\title{Nice display of the analysis}
\usage{
show_table_markdown(table)
}
\arguments{
\item{table}{The analysed data frame}

\item{...}{}
}
\description{
Nice display of the analysis
}
