% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{split_cv}
\alias{split_cv}
\title{Create n subsamples for cross validations}
\usage{
split_cv(tab, n = 10)
}
\arguments{
\item{tab}{A data.frame}

\item{n}{Number of subsamples}
}
\value{
A list of data.frames
}
\description{
Create n subsamples for cross validations
}
