% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sprintf_number_table}
\alias{sprintf_number_table}
\title{Use sprintf for displaying numbers with the right digit number}
\usage{
sprintf_number_table(string, ...)
}
\arguments{
\item{string}{the string}

\item{...}{values to be passed into string. Only logical, integer, real and character vectors are supported, but some coercion will be done: see the ‘Details’ section. Up to 100.}
}
\value{
A formatted string
}
\description{
Use sprintf for displaying numbers with the right digit number
}
