% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tab.R
\name{standardize_names}
\alias{standardize_names}
\title{Create better colnames}
\usage{
standardize_names(names, trunc = FALSE, length = 40)
}
\arguments{
\item{names}{character vector}
}
\value{
nice names vectors
}
\description{
Create better colnames
}
