% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tab.R
\name{standardize_tab}
\alias{standardize_tab}
\title{Prepare a data.frame for analysis}
\usage{
standardize_tab(tab)
}
\arguments{
\item{tab}{the data.frame}
}
\value{
A standardized data.frame
}
\description{
This function \itemize{
   \item creates a label for each columns, allowing for a better display,
   \item removes empty cells on top of the data.frame
   \item standardize the dates
   \item lowers case of levels for character vectors
   \item factors qualitative variables
   \item removes quoting characters
}
}
