% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try2}
\alias{try2}
\title{Simple tryCatch for errors and warnings}
\usage{
try2(expr, errors, warnings)
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{errors}{character vector containing error message to bypass}

\item{warnings}{character vector containing warning message to bypass}
}
\description{
Simple tryCatch for errors and warnings
}
