% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{boxplot_bivar}
\alias{boxplot_bivar}
\title{Easy Boxplot}
\usage{
boxplot_bivar(tab, x, y, palette = "hue", violin = FALSE)
}
\arguments{
\item{tab}{a data frame}

\item{x}{the independant variable}

\item{y}{the dependant variable}
}
\value{
a ggplot boxplot
}
\description{
Easy Boxplot
}
