% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-multiv.R
\name{compute_mod}
\alias{compute_mod}
\title{Compute multivariable statistical model}
\usage{
compute_mod(tab, vardep, varindep, varajust, type, pred = 0)
}
\arguments{
\item{tab}{The data.frame}

\item{vardep}{The dependant variable}

\item{varindep}{The independant variables}

\item{varajust}{The adjustment variables}

\item{type}{Type: one of "linear", "logistic" or "survival"}

\item{pred}{0 if the analysis is explanatory, 1 if the analysis is predictive, 2 if the model is computed for cross validation}
}
\value{

}
\description{
Compute multivariable statistical model
}
