% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tab.R
\name{create_tabi}
\alias{create_tabi}
\title{Create intermediate table}
\usage{
create_tabi(tab, univ)
}
\arguments{
\item{tab}{a data frame}

\item{univ}{a logical vector: TRUE if univariate analysis, FALSE otherwise}
}
\value{
a curated data frame
}
\description{
Create intermediate table
}
