% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustments.R
\name{define_varajust}
\alias{define_varajust}
\title{Detect adjustment variables}
\usage{
define_varajust(tab, vardep, varindep, type, by_lasso = TRUE, all_vars = FALSE)
}
\arguments{
\item{tab}{A data frame}

\item{vardep}{A character string of the dependant variable}

\item{varindep}{A character vector of the independant variables}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}

\item{by_lasso}{a logical: is the adjustment variable computed with lasso?}
}
\description{
Detect adjustment variables
}
