% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pval.R
\name{extract_pval}
\alias{extract_pval}
\alias{extract_pval.survdiff}
\alias{extract_pval.default}
\title{Performs univariate tests and extract p-value from objects}
\usage{
extract_pval(x, ...)

\method{extract_pval}{survdiff}(x, ...)

\method{extract_pval}{default}(x, y, ...)
}
\arguments{
\item{x}{Object of class survdiff, or a numerical or factor vector}

\item{...}{If x is a vector, this argument must be a vector of the same length}
}
\description{
Performs univariate tests and extract p-value from objects
}
