% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_env}
\alias{find_env}
\title{Find object in the parents of the calling environment}
\usage{
find_env(name, parents = rlang::current_env())
}
\arguments{
\item{name}{a length 1 character vector of the objet to find}

\item{parents}{list of environments}
}
\value{
the first environment in the list where the  object is found
}
\description{
Find object in the parents of the calling environment
}
