% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nvar_mod}
\alias{get_nvar_mod}
\title{Get the predicted number of terms in a statistical model}
\usage{
get_nvar_mod(tab, varajust)
}
\arguments{
\item{tab}{The data frame}

\item{varajust}{A character vector of adjustment variables}
}
\value{

}
\description{
Get the predicted number of terms in a statistical model
}
