% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{get_propDM}
\alias{get_propDM}
\title{Get proportion of missing data}
\usage{
get_propDM(tab)
}
\arguments{
\item{tab}{A data frame}
}
\value{
The proportion of observations having at least one missing value.
}
\description{
Get proportion of missing data
}
