% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tab.R
\name{make_tab_survival}
\alias{make_tab_survival}
\title{Add column .time}
\usage{
make_tab_survival(tab, vardep, passage = 1, dateSortie = NULL,
  dateInclusion = NULL, var_time = NULL)
}
\arguments{
\item{tab}{The data frame}

\item{vardep}{The dependent variable}

\item{passage}{If 1, The first level of the event variable is the absence of event; if 2, the second level is the absence of event}

\item{dateSortie}{Name of the column indicating event or censor date}

\item{dateInclusion}{Name of the column indicating the inclusion date}

\item{var_time}{Name of the column indicating the follow-up time follow-up}
}
\value{
A table with .time column
}
\description{
Adds the column .time, which is the follow-up time
}
\details{
This function requires either var_time or dateSortie and dateInclusion
}
