% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pourcent}
\alias{pourcent}
\title{Get the number(s) formatted in percentage}
\usage{
pourcent(nb, symbol = TRUE)
}
\arguments{
\item{nb}{a numeric vector}

\item{symbol}{displays "%"}
}
\value{
a character vector
}
\description{
Get the number(s) formatted in percentage
}
