% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustments.R
\name{prepare_varajust}
\alias{prepare_varajust}
\title{Format adjustment variables}
\usage{
prepare_varajust(tab, varajust, type)
}
\arguments{
\item{tab}{The data frame}

\item{varajust}{The adjustment variables}

\item{type}{A character string of the type of modeling, having a value among "linear", "logistic" or "survival"}
}
\value{
A character vector surrounded by "ns()" when relevant
}
\description{
Models the numeric adjustment variables with the natural spline, to be used in a formula
}
